requirements.txt
streamlit
requests
azure-ai-textanalytics
python-dotenv  # Si aún usas dotenv para cargar variables
azure-ai-language-conversations
